pkg/utils → split into two packages with meaningful names
