+++ pkg/randutil/randomhex.go
@@ -1,4 +1,4 @@
-package util
+package randutil
 
 import (
 	"crypto/rand"
