+++ pkg/rebootstrap/rebootstrap.go
@@ -1,4 +1,4 @@
-package util
+package rebootstrap
 
 import (
 	"context"
